Company manipulations now requires authentication

+++ Controller/CompanyController.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Http;
 using CompanyStructuresWebAPI.Repository;
+using CompanyStructuresWebAPI.Helper;
 
 namespace CompanyStructuresWebAPI.Interface
 {
@@ -17,34 +18,6 @@ namespace CompanyStructuresWebAPI.Interface
         public CompanyController(ICompanyRepository companyRepo)
         {
             _companyRepo = companyRepo;
-        }
-
-        bool _Check(string Username, string Password)
-        {
-            if ((Username == "KnarfRetlawReiemniets") && (Password == "MyPw"))
-                return true;
-
-            else
-                return false;
-        }
-
-        string[] _GetHeaderData(HttpRequest req)
-        {
-            string authorizationKey = Request.Headers["Authorization"].ToString();
-
-            if (authorizationKey != "")
-            {
-                authorizationKey = authorizationKey.Remove(0, 6);
-
-                string decodedKey = System.Text.Encoding.UTF8.GetString(Convert.FromBase64String(authorizationKey));
-
-                string[] authData = decodedKey.Split(":");
-
-                return authData;
-            }
-
-            else
-                return null;
         }
 
         public IActionResult GetAll()
@@ -77,9 +50,7 @@ namespace CompanyStructuresWebAPI.Interface
         [HttpPost]
         public IActionResult Post([FromBody]Model.Dto.CompanyDto company)
         {
-            string[] authData = _GetHeaderData(Request);
-
-            if ((authData != null) && (_Check(authData[0], authData[1])))
+            if (Authenticator.isAuthenticated(Request.Headers["Authorization"]))
             {
                 if (company == null)
                     return BadRequest();
@@ -100,27 +71,39 @@ namespace CompanyStructuresWebAPI.Interface
         [HttpPut]
         public IActionResult Put([FromBody] Model.Company company)
         {
-            if ((company == null) || (company.CompanyName == null))
-                return BadRequest();
+            if (Authenticator.isAuthenticated(Request.Headers["Authorization"]))
+            {
+                if (company == null)
+                    return BadRequest();
+
+                else
+                {
+                    // Exception handling
+                    _companyRepo.Update(company);
+
+                    return Created("companies", company);
+                }
+            }
 
             else
-            {
-                // Exception handling
-                _companyRepo.Update(company);
-
-                return Ok();
-            }
+                return Unauthorized();
         }
 
         [HttpDelete("{Id}")]
         public IActionResult Delete(int Id)
         {
-            // Exception handling
-            if (_companyRepo.Delete(Id) != -1)
-                return Ok();
+            if (Authenticator.isAuthenticated(Request.Headers["Authorization"]))
+            {
+                // Exception handling
+                if (_companyRepo.Delete(Id) != -1)
+                    return Ok();
+
+                else
+                    return NoContent();
+            }
 
             else
-                return NoContent();
+                return Unauthorized();
         }
     }
 }
